refactor tests - move to libcommon too
don't like that test data is now has to be imported from another lib
