+++ services/worker/tests/job_runners/split/test_descriptive_statistics.py
@@ -3,38 +3,32 @@
 from collections.abc import Callable, Mapping
 from dataclasses import replace
 from http import HTTPStatus
-from typing import Optional, Union
+from typing import Optional
 
-import numpy as np
 import pandas as pd
 import polars as pl
 import pyarrow as pa
 import pyarrow.parquet as pq
 import pytest
-from datasets import ClassLabel, Dataset
+from datasets import Dataset
 from datasets.table import embed_table_storage
 from huggingface_hub.hf_api import HfApi
 from libcommon.dtos import Priority
 from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.simple_cache import upsert_response
 from libcommon.statistics import (
-    DECIMALS,
-    MAX_NUM_STRING_LABELS,
-    MAX_PROPORTION_STRING_LABELS,
-    NO_LABEL_VALUE,
-    NUM_BINS,
     AudioColumn,
-    BoolColumn,
-    ClassLabelColumn,
     ColumnType,
-    FloatColumn,
     ImageColumn,
-    IntColumn,
-    ListColumn,
-    StringColumn,
-    generate_bins,
 )
 from libcommon.storage import StrPath
+from libs.libcommon.tests.statistics_utils import (
+    count_expected_statistics_for_bool_column,
+    count_expected_statistics_for_categorical_column,
+    count_expected_statistics_for_list_column,
+    count_expected_statistics_for_numerical_column,
+    count_expected_statistics_for_string_column,
+)
 
 from worker.config import AppConfig
 from worker.job_runners.config.parquet import ConfigParquetJobRunner
@@ -51,38 +45,6 @@ GetJobRunner = Callable[[str, str, str, AppConfig], SplitDescriptiveStatisticsJo
 GetParquetAndInfoJobRunner = Callable[[str, str, AppConfig], ConfigParquetAndInfoJobRunner]
 GetParquetJobRunner = Callable[[str, str, AppConfig], ConfigParquetJobRunner]
 GetParquetMetadataJobRunner = Callable[[str, str, AppConfig], ConfigParquetMetadataJobRunner]
-
-
-@pytest.mark.parametrize(
-    "min_value,max_value,column_type,expected_bins",
-    [
-        (0, 1, ColumnType.INT, [0, 1, 1]),
-        (0, 12, ColumnType.INT, [0, 2, 4, 6, 8, 10, 12, 12]),
-        (-10, 15, ColumnType.INT, [-10, -7, -4, -1, 2, 5, 8, 11, 14, 15]),
-        (0, 9, ColumnType.INT, [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 9]),
-        (0, 10, ColumnType.INT, [0, 2, 4, 6, 8, 10, 10]),
-        (0.0, 10.0, ColumnType.FLOAT, [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0]),
-        (0.0, 0.1, ColumnType.FLOAT, [0.0, 0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09, 0.1]),
-        (0, 0, ColumnType.INT, [0, 0]),
-        (0.0, 0.0, ColumnType.INT, [0.0, 0.0]),
-        (-0.5, 0.5, ColumnType.FLOAT, [-0.5, -0.4, -0.3, -0.2, -0.1, 0.0, 0.1, 0.2, 0.3, 0.4, 0.5]),
-        (-100.0, 100.0, ColumnType.FLOAT, [-100.0, -80.0, -60.0, -40.0, -20.0, 0.0, 20.0, 40.0, 60.0, 80.0, 100.0]),
-    ],
-)
-def test_generate_bins(
-    min_value: Union[int, float],
-    max_value: Union[int, float],
-    column_type: ColumnType,
-    expected_bins: list[Union[int, float]],
-) -> None:
-    bins = generate_bins(
-        min_value=min_value, max_value=max_value, column_name="dummy", column_type=column_type, n_bins=NUM_BINS
-    )
-    assert 2 <= len(bins) <= NUM_BINS + 1
-    if column_type is column_type.FLOAT:
-        assert pytest.approx(bins) == expected_bins
-    else:
-        assert bins == expected_bins
 
 
 @pytest.fixture
@@ -253,133 +215,6 @@ def get_parquet_metadata_job_runner(
     return _get_job_runner
 
 
-def count_expected_statistics_for_numerical_column(
-    column: pd.Series,  # type: ignore
-    dtype: ColumnType,
-) -> dict:  # type: ignore
-    minimum, maximum, mean, median, std = (
-        column.min(),
-        column.max(),
-        column.mean(),
-        column.median(),
-        column.std(),
-    )
-    n_samples = column.shape[0]
-    nan_count = column.isna().sum()
-    if nan_count == n_samples:
-        return {
-            "nan_count": n_samples,
-            "nan_proportion": 1.0,
-            "min": None,
-            "max": None,
-            "mean": None,
-            "median": None,
-            "std": None,
-            "histogram": None,
-        }
-    if dtype is ColumnType.FLOAT:
-        if minimum == maximum:
-            hist, bin_edges = np.array([column[~column.isna()].count()]), np.array([minimum, maximum])
-        else:
-            hist, bin_edges = np.histogram(column[~column.isna()], bins=NUM_BINS)
-        bin_edges = bin_edges.astype(float).round(DECIMALS).tolist()
-    else:
-        bins = generate_bins(minimum, maximum, column_name="dummy", column_type=dtype, n_bins=NUM_BINS)
-        hist, bin_edges = np.histogram(column[~column.isna()], bins)
-        bin_edges = bin_edges.astype(int).tolist()
-    hist = hist.astype(int).tolist()
-    if dtype is ColumnType.FLOAT:
-        minimum = minimum.astype(float).round(DECIMALS).item()
-        maximum = maximum.astype(float).round(DECIMALS).item()
-        mean = mean.astype(float).round(DECIMALS).item()  # type: ignore
-        median = median.astype(float).round(DECIMALS).item()  # type: ignore
-        std = std.astype(float).round(DECIMALS).item()  # type: ignore
-    else:
-        mean, median, std = list(np.round([mean, median, std], DECIMALS))
-    return {
-        "nan_count": nan_count,
-        "nan_proportion": np.round(nan_count / n_samples, DECIMALS).item() if nan_count else 0.0,
-        "min": minimum,
-        "max": maximum,
-        "mean": mean,
-        "median": median,
-        "std": std,
-        "histogram": {
-            "hist": hist,
-            "bin_edges": bin_edges,
-        },
-    }
-
-
-def count_expected_statistics_for_list_column(column: pd.Series) -> dict:  # type: ignore
-    if column.isnull().all():
-        lengths_column = pd.Series([None] * column.shape[0])
-        return count_expected_statistics_for_numerical_column(lengths_column, dtype=ColumnType.INT)
-    column_without_na = column.dropna()
-    first_sample = column_without_na.iloc[0]
-    if isinstance(first_sample, dict):  # sequence is dict of lists
-        lengths_column = column.map(lambda x: len(next(iter(x.values()))) if x is not None else None)
-    else:
-        lengths_column = column.map(lambda x: len(x) if x is not None else None)
-    return count_expected_statistics_for_numerical_column(lengths_column, dtype=ColumnType.INT)
-
-
-def count_expected_statistics_for_categorical_column(
-    column: pd.Series,  # type: ignore
-    class_label_feature: ClassLabel,
-) -> dict:  # type: ignore
-    n_samples = column.shape[0]
-    nan_count = column.isna().sum()
-    value_counts = column.value_counts(dropna=True).to_dict()
-    no_label_count = int(value_counts.pop(NO_LABEL_VALUE, 0))
-    num_classes = len(class_label_feature.names)
-    frequencies = {
-        class_label_feature.int2str(int(class_id)): value_counts.get(class_id, 0) for class_id in range(num_classes)
-    }
-    return {
-        "nan_count": nan_count,
-        "nan_proportion": np.round(nan_count / n_samples, DECIMALS).item() if nan_count else 0.0,
-        "no_label_count": no_label_count,
-        "no_label_proportion": np.round(no_label_count / n_samples, DECIMALS).item() if no_label_count else 0.0,
-        "n_unique": num_classes,
-        "frequencies": frequencies,
-    }
-
-
-def count_expected_statistics_for_string_column(column: pd.Series) -> dict:  # type: ignore
-    n_samples = column.shape[0]
-    nan_count = column.isna().sum()
-    value_counts = column.value_counts(dropna=True).to_dict()
-    n_unique = len(value_counts)
-    if (
-        n_unique / n_samples <= MAX_PROPORTION_STRING_LABELS
-        and n_unique <= MAX_NUM_STRING_LABELS
-        or n_unique <= NUM_BINS
-    ):
-        return {
-            "nan_count": nan_count,
-            "nan_proportion": np.round(nan_count / n_samples, DECIMALS).item() if nan_count else 0.0,
-            "no_label_count": 0,
-            "no_label_proportion": 0.0,
-            "n_unique": n_unique,
-            "frequencies": value_counts,
-        }
-
-    lengths_column = column.map(lambda x: len(x) if x is not None else None)
-    return count_expected_statistics_for_numerical_column(lengths_column, dtype=ColumnType.INT)
-
-
-def count_expected_statistics_for_bool_column(column: pd.Series) -> dict:  # type: ignore
-    n_samples = column.shape[0]
-    nan_count = column.isna().sum()
-    value_counts = column.value_counts(dropna=True).to_dict()
-    return {
-        "nan_count": nan_count,
-        "nan_proportion": np.round(nan_count / n_samples, DECIMALS).item() if nan_count else 0.0,
-        "frequencies": {str(key): freq for key, freq in value_counts.items()},
-    }
-
-
 @pytest.fixture
 def descriptive_statistics_expected(datasets: Mapping[str, Dataset]) -> dict:  # type: ignore
     ds = datasets["descriptive_statistics"]
@@ -540,209 +375,6 @@ def image_statistics_expected() -> dict:  # type: ignore
         },
         "partial": False,
     }
-
-
-@pytest.mark.parametrize(
-    "column_name",
-    [
-        "float__column",
-        "float__nan_column",
-        "float__all_nan_column",
-        "float__negative_column",
-        "float__cross_zero_column",
-        "float__large_values_column",
-        "float__only_one_value_column",
-        "float__only_one_value_nan_column",
-    ],
-)
-def test_float_statistics(
-    column_name: str,
-    descriptive_statistics_expected: dict,  # type: ignore
-    datasets: Mapping[str, Dataset],
-) -> None:
-    expected = descriptive_statistics_expected["statistics"][column_name]["column_statistics"]
-    data = datasets["descriptive_statistics"].to_dict()
-    computed = FloatColumn.compute_statistics(
-        data=pl.from_dict(data),
-        column_name=column_name,
-        n_samples=len(data[column_name]),
-    )
-    expected_hist, computed_hist = expected.pop("histogram"), computed.pop("histogram")
-    if computed_hist:
-        assert computed_hist["hist"] == expected_hist["hist"]
-        assert pytest.approx(computed_hist["bin_edges"]) == expected_hist["bin_edges"]
-    assert pytest.approx(computed) == expected
-    assert computed["nan_count"] == expected["nan_count"]
-
-
-@pytest.mark.parametrize(
-    "column_name",
-    [
-        "int__column",
-        "int__nan_column",
-        "int__all_nan_column",
-        "int__negative_column",
-        "int__cross_zero_column",
-        "int__large_values_column",
-        "int__only_one_value_column",
-        "int__only_one_value_nan_column",
-    ],
-)
-def test_int_statistics(
-    column_name: str,
-    descriptive_statistics_expected: dict,  # type: ignore
-    datasets: Mapping[str, Dataset],
-) -> None:
-    expected = descriptive_statistics_expected["statistics"][column_name]["column_statistics"]
-    data = datasets["descriptive_statistics"].to_dict()
-    computed = IntColumn.compute_statistics(
-        data=pl.from_dict(data),
-        column_name=column_name,
-        n_samples=len(data[column_name]),
-    )
-    print(computed)
-    expected_hist, computed_hist = expected.pop("histogram"), computed.pop("histogram")
-    if computed_hist:
-        assert computed_hist["hist"] == expected_hist["hist"]
-        assert pytest.approx(computed_hist["bin_edges"]) == expected_hist["bin_edges"]
-    assert pytest.approx(computed) == expected
-    assert computed["nan_count"] == expected["nan_count"]
-    assert computed["min"] == expected["min"]
-    assert computed["max"] == expected["max"]
-
-
-@pytest.mark.parametrize(
-    "column_name",
-    [
-        "string_text__column",
-        "string_text__nan_column",
-        "string_text__large_string_column",
-        "string_text__large_string_nan_column",
-        "string_label__column",
-        "string_label__nan_column",
-        "string_label__all_nan_column",
-    ],
-)
-def test_string_statistics(
-    column_name: str,
-    descriptive_statistics_expected: dict,  # type: ignore
-    descriptive_statistics_string_text_expected: dict,  # type: ignore
-    datasets: Mapping[str, Dataset],
-) -> None:
-    if column_name.startswith("string_text__"):
-        expected = descriptive_statistics_string_text_expected["statistics"][column_name]["column_statistics"]
-        data = datasets["descriptive_statistics_string_text"].to_dict()
-    else:
-        expected = descriptive_statistics_expected["statistics"][column_name]["column_statistics"]
-        data = datasets["descriptive_statistics"].to_dict()
-    computed = StringColumn.compute_statistics(
-        data=pl.from_dict(data),
-        column_name=column_name,
-        n_samples=len(data[column_name]),
-    )
-    if column_name.startswith("string_text__"):
-        expected_hist, computed_hist = expected.pop("histogram"), computed.pop("histogram")
-        assert expected_hist["hist"] == computed_hist["hist"]
-        assert expected_hist["bin_edges"] == pytest.approx(computed_hist["bin_edges"])
-        assert expected == pytest.approx(computed)
-    else:
-        assert expected == computed
-
-
-@pytest.mark.parametrize(
-    "column_name",
-    [
-        "class_label__column",
-        "class_label__nan_column",
-        "class_label__all_nan_column",
-        "class_label__less_classes_column",
-        "class_label__string_column",
-        "class_label__string_nan_column",
-        "class_label__string_all_nan_column",
-    ],
-)
-def test_class_label_statistics(
-    column_name: str,
-    descriptive_statistics_expected: dict,  # type: ignore
-    datasets: Mapping[str, Dataset],
-) -> None:
-    expected = descriptive_statistics_expected["statistics"][column_name]["column_statistics"]
-    class_label_feature = datasets["descriptive_statistics"].features[column_name]
-    data = datasets["descriptive_statistics"].to_dict()
-    computed = ClassLabelColumn.compute_statistics(
-        data=pl.from_dict(data),
-        column_name=column_name,
-        n_samples=len(data[column_name]),
-        feature_dict={"_type": "ClassLabel", "names": class_label_feature.names},
-    )
-    assert expected == computed
-
-
-@pytest.mark.parametrize(
-    "column_name",
-    [
-        "bool__column",
-        "bool__nan_column",
-        "bool__all_nan_column",
-    ],
-)
-def test_bool_statistics(
-    column_name: str,
-    descriptive_statistics_expected: dict,  # type: ignore
-    datasets: Mapping[str, Dataset],
-) -> None:
-    expected = descriptive_statistics_expected["statistics"][column_name]["column_statistics"]
-    data = datasets["descriptive_statistics"].to_dict()
-    computed = BoolColumn.compute_statistics(
-        data=pl.from_dict(data),
-        column_name=column_name,
-        n_samples=len(data[column_name]),
-    )
-    assert computed == expected
-
-
-@pytest.mark.parametrize(
-    "column_name",
-    [
-        "list__int_column",
-        "list__int_nan_column",
-        "list__int_all_nan_column",
-        "list__string_column",
-        "list__string_nan_column",
-        "list__string_all_nan_column",
-        "list__dict_column",
-        "list__dict_nan_column",
-        "list__dict_all_nan_column",
-        "list__sequence_int_column",
-        "list__sequence_int_nan_column",
-        "list__sequence_int_all_nan_column",
-        "list__sequence_class_label_column",
-        "list__sequence_class_label_nan_column",
-        "list__sequence_class_label_all_nan_column",
-        "list__sequence_of_sequence_bool_column",
-        "list__sequence_of_sequence_bool_nan_column",
-        "list__sequence_of_sequence_bool_all_nan_column",
-        "list__sequence_of_sequence_dict_column",
-        "list__sequence_of_sequence_dict_nan_column",
-        "list__sequence_of_sequence_dict_all_nan_column",
-        "list__sequence_of_list_dict_column",
-        "list__sequence_of_list_dict_nan_column",
-        "list__sequence_of_list_dict_all_nan_column",
-    ],
-)
-def test_list_statistics(
-    column_name: str,
-    descriptive_statistics_expected: dict,  # type: ignore
-    datasets: Mapping[str, Dataset],
-) -> None:
-    expected = descriptive_statistics_expected["statistics"][column_name]["column_statistics"]
-    data = datasets["descriptive_statistics"].to_dict()
-    computed = ListColumn.compute_statistics(
-        data=pl.from_dict(data),
-        column_name=column_name,
-        n_samples=len(data[column_name]),
-    )
-    assert computed == expected
 
 
 @pytest.fixture
